requirements.txt
google-generativeai
sentence-transformers
nltk
torch
codeswitch
google-cloud-translate
googletrans